api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.36
asyncpg==0.30.0
python-multipart==0.0.12
websockets==13.1
httpx==0.27.2
